requirements.txt
PyQt6==6.4.0
opencv-python==4.7.0.72
numpy==1.24.3
Pillow==9.5.0
torch==2.0.1
transformers==4.30.2
easyocr==1.7.0
openai==1.0.0
python-dotenv==1.0.0
pyttsx3==2.90
requests==2.31.0
tqdm==4.65.0
pyinstaller==6.3.0 